requirements.txt
# ML Pipeline
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0

# Frontend
streamlit>=1.35.0
plotly>=5.18.0

# Auth
bcrypt>=4.0.0

# Email alerts (built-in smtplib used — no extra package needed)

# Optional: load env vars from .env file
python-dotenv>=1.0.0